requirements_test.txt
pytest-homeassistant-custom-component>=0.13.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
